+++ login-workflow/src/components/password/ChangePasswordForm.tsx
@@ -1,11 +1,13 @@
 import React, { ChangeEvent, useState, useCallback, MutableRefObject } from 'react';
-import { useLanguageLocale } from '@brightlayer-ui/react-auth-shared';
+import { useInjectedUIContext, useLanguageLocale } from '@brightlayer-ui/react-auth-shared';
 import Typography from '@mui/material/Typography';
 import Divider from '@mui/material/Divider';
 import { useTheme } from '@mui/material/styles';
 import { SecureTextField } from '../SecureTextField';
 import { PasswordRequirements } from './PasswordRequirements';
 import { FullDividerStyles, TextFieldStyles } from '../../styles';
+import CheckCircleOutlinedIcon from '@mui/icons-material/CheckCircleOutlined';
+import { defaultPasswordRequirements } from '../../constants';
 
 export type ChangePasswordFormProps = {
     onPasswordChange: (passwords: { password: string; confirm: string }) => void;
@@ -56,6 +58,8 @@ export const ChangePasswordForm: React.FC<React.PropsWithChildren<React.PropsWit
     // Local State
     const [passwordInput, setPasswordInput] = useState(initialPassword);
     const [confirmInput, setConfirmInput] = useState(initialConfirm);
+    const [shouldValidateConfirmPassword, setShouldValidateConfirmPassword] = useState(false);
+    const [shouldValidatePassword, setShouldValidatePassword] = useState(false);
 
     const onPassChange = useCallback(
         (newPassword: any) => {
@@ -72,6 +76,20 @@ export const ChangePasswordForm: React.FC<React.PropsWithChildren<React.PropsWit
         },
         [setConfirmInput, onPasswordChange, passwordInput]
     );
+
+    const hasConfirmPasswordError = useCallback(
+        (): boolean => shouldValidateConfirmPassword && confirmInput.length !== 0 && confirmInput !== passwordInput,
+        [shouldValidateConfirmPassword, confirmInput, passwordInput]
+    );
+
+    const { passwordRequirements = defaultPasswordRequirements(t) } = useInjectedUIContext();
+
+    const isValidPassword = useCallback((): boolean => {
+        for (let i = 0; i < passwordRequirements.length; i++) {
+            if (!new RegExp(passwordRequirements[i].regex).test(passwordInput)) return false;
+        }
+        return true;
+    }, [passwordRequirements, passwordInput]);
 
     return (
         <>
@@ -93,6 +111,8 @@ export const ChangePasswordForm: React.FC<React.PropsWithChildren<React.PropsWit
                         confirmRef.current.focus();
                     }
                 }}
+                error={shouldValidatePassword && !isValidPassword()}
+                onBlur={(): void => setShouldValidatePassword(true)}
             />
             <PasswordRequirements sx={{ mt: 2 }} passwordText={passwordInput} />
             <SecureTextField
@@ -106,6 +126,14 @@ export const ChangePasswordForm: React.FC<React.PropsWithChildren<React.PropsWit
                 onKeyPress={(e): void => {
                     if (e.key === 'Enter' && onSubmit) onSubmit();
                 }}
+                error={hasConfirmPasswordError()}
+                helperText={hasConfirmPasswordError() ? t('blui:FORMS.PASS_MATCH_ERROR') : ''}
+                icon={
+                    confirmInput.length !== 0 && confirmInput === passwordInput ? (
+                        <CheckCircleOutlinedIcon color="success" />
+                    ) : undefined
+                }
+                onBlur={(): void => setShouldValidateConfirmPassword(true)}
             />
         </>
     );